Use a theme-agnostic way to add data-incom-comment attributes to comments.
Themes that render comment sections using WP's comments block
do not respect legacy wp_list_comments() modifications. Thus,
the inline-comments legacy method for adding the attribute did not
work in themes like Twenty Twenty-Four. As a result, comments were
not correctly associated with their text elements.

As a workaround, we build an index of comments and associated
text elements as a JavaScript object. This object is then used
to add the data-incom-comment attribute to comments, in the client.